Avoid circular references in core.ts
